+++ main.py
@@ -1,8 +1,8 @@
-from crawler.main import main as crawler_main
-from crawler.m_configs import MERCARI_IT_BOOK_URL, QUANTITY
+from scrapper.main import main as scrapper_main
+from scrapper.m_configs import MERCARI_IT_BOOK_URL, QUANTITY
 
 def main():     
-    crawler_main(main_url=MERCARI_IT_BOOK_URL, quantity=QUANTITY)
+    scrapper_main(main_url=MERCARI_IT_BOOK_URL, quantity=QUANTITY)
 
 
 if __name__ == '__main__':
